Fixed bug where move would not work for some angle values

+++ Unknown 1.0++/Sprite.h
@@ -16,7 +16,7 @@ namespace Unknown
 		double angle;
 
 	public:
-		Point<int> location;
+		Point<double> location;
 		Vector direction;
 		mutable AABB bounds;
